Extracted simulation logic in separate service, fixed bug with incorrect speed, added speed meter scale, extracted traffic generator factor for each road
