+++ readme.md
@@ -1,0 +1,3 @@
+#Road Modeller
+This application for traffic simulation. Now supports traffic lights time and speed limit optimization tasks. You can use default road state or create custom scenario.
+Using Spring Reactive WebSockets and state saving in MongoDB.
