+++ src/main/java/com/drakezzz/roadmodeller/model/processor/SimulationProcessor.java
@@ -1,0 +1,9 @@
+package com.drakezzz.roadmodeller.model.processor;
+
+import com.drakezzz.roadmodeller.persistence.entity.ModelState;
+
+public interface SimulationProcessor {
+
+    ModelState simulate(ModelState modelState);
+
+}
